Rename ConfigForUpdate to Config::ToConfigure
